perlin noise generation

+++ htw_core_collections.c
@@ -4,7 +4,7 @@
 
 #define GET_ITEM_POINTER(list, index) (list->items + (index * list->itemSize))
 
-List *createList (int itemSize, int initialLength) {
+List *createList(size_t itemSize, u32 initialLength) {
     void *items = malloc(itemSize * initialLength);
     // malloc l instead of using struct assignment to avoid returning the address of a local variable
     List *l = (List*)malloc(sizeof(List));
@@ -16,7 +16,7 @@ List *createList (int itemSize, int initialLength) {
 }
 
 // TODO: do something to indicate an error when out of range
-List *sliceList(List *originalList, int startIndex, int sliceLength) {
+List *sliceList(List *originalList, u32 startIndex, u32 sliceLength) {
     if (startIndex >= originalList->length || startIndex + sliceLength > originalList->length) {
         return NULL;
     }
@@ -45,17 +45,19 @@ void destroyList(List *list) {
 
 // TODO: add out of range check
 // return a pointer to the item at the given index
-void *getItem(List *list, int index) {
+void *getItem(List *list, u32 index)
+{
     return GET_ITEM_POINTER(list, index);
 }
 
 // Uses memcpy to add the value of *newItem to list[index]
-void setItem(List *list, int index, void *newItem) {
+void setItem(List *list, u32 index, void *newItem) {
     memcpy(GET_ITEM_POINTER(list, index), newItem, list->itemSize);
 }
 
 // Returns the index of the newly added item
-int pushItem(List *list, void *newItem) {
+u32 pushItem(List *list, void *newItem)
+{
     if (list->length == list->capacity)
         addCapacity(list);
     memcpy(GET_ITEM_POINTER(list, list->length), newItem, list->itemSize);
